refatcor: move components out organisms
