+++ packages/shared/components/interfaces/node.interface.ts
@@ -1,0 +1,6 @@
+export interface NodeValidationOptions {
+    checkNodeInfo: boolean
+    checkSameNetwork: boolean
+    uniqueCheck: boolean
+    validateClientOptions: boolean
+}
