+++ packages/shared/components/organisms/index.js
@@ -1,3 +1,1 @@
 export { default as AccountAssetsList } from './AccountAssetsList.svelte'
-export { default as NodeConfigurationForm } from './NodeConfigurationForm.svelte'
-export { default as NodeListTable } from './NodeListTable.svelte'
